+++ tests/test_providers.py
@@ -285,3 +285,16 @@ class TestProviders:
 
         finally:
             signal.signal(SIGINT, old_handler)
+
+
+def test_provider_cache():
+    injector = get_injector_for_testing({})
+    request = get_response_for_testing(None).request
+    provider = AutoExtractProvider(injector.crawler)
+    fingerprint = provider.fingerprint(
+        {AutoExtractHtml, AutoExtractProductData, AutoExtractArticleData},
+        request)
+    assert fingerprint == """{"article": {"articleBodyRaw": false, "url": "http://example.com"}, "product": {"articleBodyRaw": false, "fullHtml": true, "url": "http://example.com"}}"""
+    assert provider.serialize("foo") == "foo"
+    assert provider.deserialize("bar") == "bar"
+    assert provider.has_cache_support