+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 TBR
 ------------------
 
+* The AutoExtract provider now supports the new providers API for caching
+* The preferred pageType for HTML requests (``AutoExtractProductData``)
+  is now chosen always if listed as dependency instead of just choosing
+  the first dependency ``pageType`` to request the HTML
 * removed support for Python 3.6
 * added support for Python 3.10